Fix package name problem
The mandatory ‘Package’ field gives the name of the package.
This should contain only (ASCII) letters, numbers and dot, have at least two
characters and start with a letter and not end in a dot.
If it needs explaining, this should be done in the ‘Description’
field (and not the ‘Title’ field).